fix code conflict from book card to usewishlist
